Quiescence fixed and tweaked. Now on JENCE level

+++ src/engine/search/mod.rs
@@ -33,9 +33,9 @@ pub const MATE_VALUE: i16 = 29000;
 pub const MATE_BOUND: i16 = 28000;
 
 /// Initial aspiration window is +- this value
-pub const ASPIRATION_WINDOW: i16 = 15;
+pub const ASPIRATION_WINDOW: i32 = 15;
 /// Exponentially increase window by this multiplier on fail
-pub const ASPIRATION_WINDOW_MULT: i16 = 3;
+pub const ASPIRATION_WINDOW_MULT: i32 = 3;
 
 /// Amount of killer moves pr. ply to remember. Must be at least 1
 pub const KILLER_MOVE_COUNT: usize = 3;
